+++ src/Forms/UI/views/previews/fields/hidden.blade.php
@@ -1,1 +1,3 @@
-{{-- like no way we'll show our hidden field in a window dog --}}
+{{-- Like no way we'll show our hidden field in a window dog --}}
+{{-- That's why we render an empty div here, so any preview container can be hidden too if it contains a hidden field --}}
+<div data-slot="hidden-field-preview"></div>