+++ src/Table/UI/views/columns/image.blade.php
@@ -1,13 +1,12 @@
 @foreach ($getItems() as $item)
-    <div class="flex size-8 items-center justify-center overflow-hidden rounded-md bg-grey-100">
-
+    <div class="bg-grey-100 flex size-8 items-center justify-center overflow-hidden rounded-md">
         @if ($item->hasLink())
             <a
                 href="{{ $item->getLink() }}"
                 title="{{ $item->getValue($getLocale()) }}"
                 {!! $item->shouldOpenInNewTab() ? 'target="_blank" rel="noopener"' : '' !!}
                 {!! $item->getCustomAttributesAsString() !!}
-                class="leading-5 text-grey-800 hover:underline hover:underline-offset-2"
+                class="text-grey-800 leading-5 hover:underline hover:underline-offset-2"
             >
                 @if ($item->getValue($getLocale()))
                     <img src="{{ $item->getValue($getLocale()) }}" class="h-full w-full object-cover" />
@@ -15,12 +14,14 @@
             </a>
         @else
             @if ($item->getValue($getLocale()))
-                <img src="{{ $item->getValue($getLocale()) }}"
-                     class="h-full w-full object-cover" {!! $item->getCustomAttributesAsString() !!} />
+                <img
+                    src="{{ $item->getValue($getLocale()) }}"
+                    class="h-full w-full object-cover"
+                    {!! $item->getCustomAttributesAsString() !!}
+                />
             @else
-                <x-chief::icon.image class="size-5 text-grey-500" />
+                <x-chief::icon.image class="text-grey-500 size-5" />
             @endif
         @endif
     </div>
-
 @endforeach