+++ CHANGELOG.md
@@ -5,6 +5,7 @@ the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
 ## Unreleased
 
+- Add: Window Layout Component
 - Changed: AdminToast element layout update
 
 ## [0.10.12] - 2025-09-10