+++ CHANGELOG.md
@@ -6,6 +6,7 @@ the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 ## Unreleased
 
 - Add: Window Layout Component
+- Add: Basic Table Components. These are used by the Livewire Table Component for displaying default data. The basic components can be used to build a custom table with the same layout.
 - Changed: AdminToast element layout update
 
 ## [0.10.12] - 2025-09-10